requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.3
fastapi==0.109.2
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.4
idna==3.6
multidict==6.0.5
pycparser==2.21
pydantic==2.6.1
pydantic_core==2.16.2
PyJWT==2.8.0
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.31.0
sniffio==1.3.0
starlette==0.36.3
typing_extensions==4.9.0
urllib3==2.2.1
uvicorn==0.27.1
yarl==1.9.4
